[MOL-16700][RL] Update LocalNav documentation
