--- stories/local-nav/local-nav-menu/local-nav-menu.mdx
@@ -1,22 +1,0 @@
-import { Canvas, Meta } from "@storybook/blocks";
-import * as LocalNavMenuStories from "./local-nav-menu.stories";
-import { PropsTable } from "./props-table";
-import { Heading3, Secondary, Title } from "../../storybook-common";
-
-<Meta of={LocalNavMenuStories} />
-
-<Title>NavMenu</Title>
-
-<Secondary>Overview</Secondary>
-
-Menu Component for mobile scenarios where it should be sticky when it reaches the top of the screen.
-
-```tsx
-import { LocalNavMenu } from "@lifesg/react-design-system/local-nav/LocalNavMenu";
-```
-
-<Canvas of={LocalNavMenuStories.Default} />
-
-<Secondary>Component API</Secondary>
-
-<PropsTable />
